Move ControlPlaneApi and related types up into control_plane module.
